+++ src/components/organisms/SideNav/SideNav.styles.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+  background-color: ${({ theme }) => theme.colors.darkGray};
+  grid-row: 2 / 3;
+  grid-column: 1 / 2;
+  display: grid;
+  grid-template-rows: repeat(4, 4em);
+  padding-top: 1em;
+  gap: 1em;
+`;
