+++ src/views/Words-list/WordsList.js
@@ -1,10 +1,10 @@
 import React, { useContext, useState } from 'react';
-import ListItem from '../../molecules/ListItem/ListItem';
-import Modal from '../../molecules/Modal/Modal';
-import SearchBar from '../../molecules/SearchBar/SearchBar';
+import ListItem from '../../components/molecules/ListItem/ListItem';
+import Modal from '../../components/molecules/Modal/Modal';
+import SearchBar from '../../components/molecules/SearchBar/SearchBar';
+import ConfirmModal from '../../components/molecules/ConfirmModal/ConfirmModal';
 import { StyledList } from './WordsList.styles';
-import { StorageContext } from '../../../StorageProvider';
-import ConfirmModal from '../../molecules/ConfirmModal/ConfirmModal';
+import { StorageContext } from '../../StorageProvider';
 
 const WordsList = () => {
   const { handleDelete, filtered, phrase } = useContext(StorageContext);
